+++ src/shortestpaths/shortestpaths.jl
@@ -1,0 +1,6 @@
+abstract AbstractPathState
+
+include("astar.jl")
+include("bellman-ford.jl")
+include("dijkstra.jl")
+include("floyd-warshall.jl")